TIP-860: Move Product related api tests
