+++ tests/back/Pim/Enrichment/EndToEnd/Product/Product/ExternalApi/ErrorListProductEndToEnd.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\Product;
+namespace AkeneoTest\Pim\Enrichment\EndToEnd\Product\Product\ExternalApi;
 
 use Akeneo\Test\Integration\Configuration;
 use Symfony\Bundle\FrameworkBundle\Client;
 use Symfony\Component\HttpFoundation\Response;
 
-class ErrorListProductIntegration extends AbstractProductTestCase
+class ErrorListProductEndToEnd extends AbstractProductTestCase
 {
     public function testNotFoundChannel()
     {
