+++ tests/back/Pim/Enrichment/EndToEnd/Product/Product/ExternalApi/ListProductWithCompletenessEndToEnd.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\Product;
+namespace AkeneoTest\Pim\Enrichment\EndToEnd\Product\Product\ExternalApi;
 
 use Akeneo\Test\Integration\Configuration;
 use Doctrine\Common\Collections\Collection;
@@ -10,7 +10,7 @@ use Doctrine\Common\Collections\Collection;
 /**
  * @group ce
  */
-class ListProductWithCompletenessIntegration extends AbstractProductTestCase
+class ListProductWithCompletenessEndToEnd extends AbstractProductTestCase
 {
     /** @var Collection */
     private $products;
