+++ tests/back/Pim/Enrichment/EndToEnd/Product/Product/ExternalApi/SuccessListProductEndToEnd.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\Product;
+namespace AkeneoTest\Pim\Enrichment\EndToEnd\Product\Product\ExternalApi;
 
 use Akeneo\Pim\Enrichment\Component\Product\Query\Filter\Operators;
 use Akeneo\Test\Integration\Configuration;
@@ -11,7 +11,7 @@ use Doctrine\Common\Collections\Collection;
 /**
  * @group ce
  */
-class SuccessListProductIntegration extends AbstractProductTestCase
+class SuccessListProductEndToEnd extends AbstractProductTestCase
 {
     /** @var Collection */
     private $products;
